Greptile and CoPilot changes
